# Remove char_indices to reduce indirectrion