* Fix one bug in cut_all, where single char should be counted since they
are in the dictionary with freq > 0


master
```
Gnuplot not found, disabling plotting
jieba cut no hmm        time:   [7.7392 us 7.7781 us 7.8221 us]
                        change: [-6.4848% -5.3074% -4.1575%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) high mild
  6 (6.00%) high severe

jieba cut with hmm      time:   [10.067 us 10.133 us 10.221 us]
                        change: [+1.0484% +2.3710% +3.6592%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 13 outliers among 100 measurements (13.00%)
  5 (5.00%) high mild
  8 (8.00%) high severe
```

this branch
```
Gnuplot not found, disabling plotting
jieba cut no hmm        time:   [7.2088 us 7.2327 us 7.2618 us]
                        change: [-15.431% -14.305% -13.246%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  2 (2.00%) low mild
  6 (6.00%) high mild
  7 (7.00%) high severe

jieba cut with hmm      time:   [9.6546 us 9.7062 us 9.7624 us]
                        change: [-17.284% -15.481% -13.722%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  3 (3.00%) high mild
  9 (9.00%) high severe

Benchmarking jieba cut_all:
```

master
```
➜  jieba-rs git:(master) ./target/release/weicheng
4463ms
```

this branch
```
➜  jieba-rs git:(remove-char-indices) ./target/release/weicheng
4330ms
```